started adding base class functionality

+++ class_tmpl/create.class.sh
@@ -53,16 +53,22 @@ then
 			rm Makefile.tmp2
 		fi
 		PRINT_DEBUG ${INPUT_PATH:-"~/bin"}
-		cat ${INPUT_PATH:-"/home/brian/bin"}/class.hpp.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}/g" > ${OUTPUT_PATH:-".."}/src/${CLASS_NAME}.hpp
-		cat ${INPUT_PATH:-"/home/brian/bin"}/class.cpp.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}/g" > ${OUTPUT_PATH:-".."}/src/${CLASS_NAME}.cpp	
-		cat Makefile.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}.o  @@CLASS_NAME@@/g" > Makefile.tmp
-		# try to update Makefile with new rule
-		MAKE_RULE=$(cat ${INPUT_PATH:-"/home/brian/bin"}/make.class.snip.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}/g")
-		# make a backup of Makefile for now
-		cat Makefile.tmp | sed "s/## auto gernerated here ##/${MAKE_RULE}\n## auto gernerated here ##/g" > Makefile.tmpl
-		cat Makefile.tmpl | sed "s/@@CLASS_NAME@@//g" > Makefile
-		rm  Makefile.tmp
-		popd
+		
+		if [ ! -z $CLASS_NAME ]; then
+
+			cat ${INPUT_PATH:-"/home/brian/bin"}/class.hpp.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}/g" > ${OUTPUT_PATH:-".."}/src/${CLASS_NAME}.hpp
+			cat ${INPUT_PATH:-"/home/brian/bin"}/class.cpp.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}/g" > ${OUTPUT_PATH:-".."}/src/${CLASS_NAME}.cpp	
+			cat Makefile.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}.o  @@CLASS_NAME@@/g" > Makefile.tmp
+			# try to update Makefile with new rule
+			MAKE_RULE=$(cat ${INPUT_PATH:-"/home/brian/bin"}/make.class.snip.tmpl | sed "s/@@CLASS_NAME@@/${CLASS_NAME}/g")
+			# make a backup of Makefile for now
+			cat Makefile.tmp | sed "s/## auto gernerated here ##/${MAKE_RULE}\n## auto gernerated here ##/g" > Makefile.tmpl
+			cat Makefile.tmpl | sed "s/@@CLASS_NAME@@//g" > Makefile
+			rm  Makefile.tmp
+			popd
+		else
+			echo Test
+		fi
 
 	else
 		echo "Error: This is not a project directory."
